i_slint_backend_winit

Struct BackendBuilder

pub struct BackendBuilder { /* private fields */ }

Use the BackendBuilder to configure the properties of the Winit Backend before creating it. Create the builder using Backend::builder(), then configure it for example with Self::with_renderer_name, and build the backend using Self::build.

Implementations

impl BackendBuilder

pub fn request_graphics_api(self, graphics_api: RequestedGraphicsAPI) -> Self

Configures this builder to require a renderer that supports the specified graphics API.

pub fn with_renderer_name(self, name: impl Into<String>) -> Self

Configures this builder to use the specified renderer name when building the backend later. Pass renderer-software for example to configure the backend to use the Slint software renderer.

pub fn with_window_attributes_hook( self, hook: impl Fn(WindowAttributes) -> WindowAttributes + 'static, ) -> Self

Configures this builder to use the specified hook that will be called before a Window is created.

It can be used to adjust settings of window that will be created.

Example

let mut backend = i_slint_backend_winit::Backend::builder()
    .with_window_attributes_hook(|attributes| attributes.with_content_protected(true))
    .build()
    .unwrap();
slint::platform::set_platform(Box::new(backend));

pub fn with_event_loop_builder( self, event_loop_builder: EventLoopBuilder<SlintUserEvent>, ) -> Self

Configures this builder to use the specified event loop builder when creating the event loop during a subsequent call to Self::build.

pub fn build(self) -> Result<Backend, PlatformError>

Builds the backend with the parameters configured previously. Set the resulting backend with slint::platform::set_platform():

Example

let mut backend = i_slint_backend_winit::Backend::builder()
    .with_renderer_name("renderer-software")
    .build()
    .unwrap();
slint::platform::set_platform(Box::new(backend));